requirements.txt
azure-functions
    # Required for the Azure Functions runtime
    azure-functions-worker